Clean up attributes
clean up all the attributes with wrong product class if product class is switched
when using patch, but if there are also product attributes with the same code
on the new productclass, then the attributevalue must be chenged to that attribute
on the new product class.

This is because you will get very very buggy operation of product.attr.something
if you don't do that.